+++ python/dolfin/__init__.py
@@ -32,11 +32,12 @@ del sys
 
 # Import cpp modules
 from .cpp.common import (Variable, has_debug, has_hdf5, has_scotch,
-                         has_hdf5_parallel, has_mpi, has_petsc,
-                         has_petsc4py, has_parmetis, has_slepc,
-                         has_slepc4py, git_commit_hash, DOLFIN_EPS,
-                         DOLFIN_PI, TimingClear, TimingType, timing,
-                         timings, list_timings, dump_timings_to_xml)
+                         has_hdf5_parallel, has_mpi, has_mpi4py,
+                         has_petsc, has_petsc4py, has_parmetis,
+                         has_slepc, has_slepc4py, git_commit_hash,
+                         DOLFIN_EPS, DOLFIN_PI, TimingClear, TimingType,
+                         timing, timings, list_timings, dump_timings_to_xml,
+                         SubSystemsManager)
 
 if has_hdf5():
     from .cpp.adaptivity import TimeSeries
@@ -138,10 +139,11 @@ from .function.function import Function
 from .function.argument import (TestFunction, TrialFunction,
                                 TestFunctions, TrialFunctions)
 from .function.constant import Constant
-from .function.specialfunctions import (FacetNormal, CellSize,
-                                        SpatialCoordinate, CellVolume,
-                                        Circumradius, FacetArea,
-                                        MeshCoordinates)
+from .function.specialfunctions import (MeshCoordinates, FacetArea, FacetNormal,
+                                        CellVolume, SpatialCoordinate, CellNormal,
+                                        CellDiameter, Circumradius,
+                                        MinCellEdgeLength, MaxCellEdgeLength,
+                                        MinFacetEdgeLength, MaxFacetEdgeLength)
 from .function.expression import Expression, UserExpression, CompiledExpression
 
 # experimental