+++ python/src/mesh.cpp
@@ -115,8 +115,10 @@ namespace dolfin_wrappers
       .def(py::init<>())
       .def(py::init<std::string>())
       .def(py::init<const dolfin::Mesh&>())
-      .def(py::init<MPI_Comm>())
-      .def(py::init<MPI_Comm, std::string>())  // Put MPI constructors last to avoid casting problems
+      .def(py::init([](const MPICommWrapper comm)
+                    { return std::unique_ptr<dolfin::Mesh>(new dolfin::Mesh(comm.get())); }))
+      .def(py::init([](const MPICommWrapper comm, const std::string filename)
+                    { return std::unique_ptr<dolfin::Mesh>(new dolfin::Mesh(comm.get(), filename)); }))
       .def("bounding_box_tree", &dolfin::Mesh::bounding_box_tree)
       .def("cells", [](const dolfin::Mesh& self)
            {
@@ -142,6 +144,7 @@ namespace dolfin_wrappers
            &dolfin::Mesh::data, "Data associated with a mesh")
       .def("geometry", (dolfin::MeshGeometry& (dolfin::Mesh::*)()) &dolfin::Mesh::geometry,
            py::return_value_policy::reference, "Mesh geometry")
+      .def("hash", &dolfin::Mesh::hash)
       .def("hmax", &dolfin::Mesh::hmax)
       .def("hmin", &dolfin::Mesh::hmin)
       .def("id", &dolfin::Mesh::id)
@@ -155,8 +158,10 @@ namespace dolfin_wrappers
              auto _o = o.attr("_cpp_object").cast<dolfin::Expression*>();
              self.init_cell_orientations(*_o);
            })
-      .def("mpi_comm", &dolfin::Mesh::mpi_comm)
-      .def("num_entities", &dolfin::Mesh::num_entities, "Number of mesh entities")
+      .def("mpi_comm", [](dolfin::Mesh& self)
+           { return MPICommWrapper(self.mpi_comm()); })
+      .def("num_entities", &dolfin::Mesh::num_entities,
+           "Number of mesh entities")
       .def("num_vertices", &dolfin::Mesh::num_vertices, "Number of vertices")
       .def("num_edges", &dolfin::Mesh::num_edges, "Number of edges")
       .def("num_faces", &dolfin::Mesh::num_faces, "Number of faces")
@@ -173,7 +178,8 @@ namespace dolfin_wrappers
       .def("size_global", &dolfin::Mesh::size_global)
       .def("smooth", &dolfin::Mesh::smooth, py::arg("num_iterations")=1)
       .def("smooth_boundary", &dolfin::Mesh::smooth_boundary)
-      .def("snap_boundary", &dolfin::Mesh::snap_boundary, py::arg("subdomain"), py::arg("harmonic_smoothing")=true)
+      .def("snap_boundary", &dolfin::Mesh::snap_boundary, py::arg("subdomain"),
+           py::arg("harmonic_smoothing")=true)
       .def("topology", (const dolfin::MeshTopology& (dolfin::Mesh::*)() const)
            &dolfin::Mesh::topology, "Mesh topology")
       .def("translate", &dolfin::Mesh::translate)
@@ -209,9 +215,9 @@ namespace dolfin_wrappers
     // dolfin::MeshConnectivity class
     py::class_<dolfin::MeshConnectivity, std::shared_ptr<dolfin::MeshConnectivity>>
       (m, "MeshConnectivity", "DOLFIN MeshConnectivity object")
-      .def("__call__", [](const dolfin::MeshConnectivity& self, std::size_t i){
-          return Eigen::Map<const Eigen::Matrix<unsigned int, Eigen::Dynamic, 1>>
-            (self(i), self.size(i));})
+      .def("__call__", [](const dolfin::MeshConnectivity& self, std::size_t i)
+           {
+             return Eigen::Map<const Eigen::Matrix<unsigned int, Eigen::Dynamic, 1>>(self(i), self.size(i));})
       .def("size", (std::size_t (dolfin::MeshConnectivity::*)() const)
            &dolfin::MeshConnectivity::size)
       .def("size", (std::size_t (dolfin::MeshConnectivity::*)(std::size_t) const)
@@ -231,10 +237,10 @@ namespace dolfin_wrappers
       .def("num_global_entities", &dolfin::MeshEntity::num_global_entities,
            "Global number of incident entities of given dimension")
       .def("entities", [](dolfin::MeshEntity& self, std::size_t dim)
-        {
-          return Eigen::Map<const Eigen::Matrix<unsigned int, Eigen::Dynamic, 1>>
-            (self.entities(dim), self.num_entities(dim));
-        })
+           {
+             return Eigen::Map<const Eigen::Matrix<unsigned int, Eigen::Dynamic, 1>>
+               (self.entities(dim), self.num_entities(dim));
+           })
       .def("midpoint", &dolfin::MeshEntity::midpoint, "Midpoint of Entity")
       .def("sharing_processes", &dolfin::MeshEntity::sharing_processes)
       .def("is_shared", &dolfin::MeshEntity::is_shared)
@@ -317,7 +323,6 @@ namespace dolfin_wrappers
              return *self;
            });
 
-
     m.def("entities", [](dolfin::Mesh& mesh, std::size_t dim)
           { return dolfin::MeshEntityIterator(mesh, dim); });
     m.def("entities", [](dolfin::MeshEntity& meshentity, std::size_t dim)
@@ -378,6 +383,7 @@ namespace dolfin_wrappers
       .def("id", &dolfin::MeshFunction<SCALAR>::id) \
       .def("ufl_id", &dolfin::MeshFunction<SCALAR>::id) \
       .def("mesh", &dolfin::MeshFunction<SCALAR>::mesh) \
+      .def("set_values", &dolfin::MeshFunction<SCALAR>::set_values) \
       .def("set_all", &dolfin::MeshFunction<SCALAR>::set_all) \
       .def("where_equal", &dolfin::MeshFunction<SCALAR>::where_equal) \
       .def("array", [](dolfin::MeshFunction<SCALAR>& self) \