+++ python/src/petsc_casters.h
@@ -27,166 +27,81 @@
 #include <petscvec.h>
 #include <petscdm.h>
 
+// pybind11 casters for PETSc/petsc4py objects
 #ifdef HAS_PYBIND11_PETSC4PY
 #include <petsc4py/petsc4py.h>
+
+// Import petsc4py on demand
+#define VERIFY_PETSC4PY(func)   \
+  if (!func)                    \
+  {                             \
+    if (import_petsc4py() != 0) \
+    {                           \
+      std::cout << "ERROR: could not import petsc4py!" << std::endl; \
+      throw std::runtime_error("Error when importing petsc4py");     \
+    }                           \
+  }
+
+// Macro for casting between dolfin and petsc4py objects
+#define PETSC_CASTER_MACRO(TYPE, NAME)          \
+  template <> class type_caster<_p_##TYPE>      \
+    {                                           \
+    public:                                     \
+      PYBIND11_TYPE_CASTER(TYPE, _(#NAME));     \
+      bool load(handle src, bool)               \
+      {                                         \
+        VERIFY_PETSC4PY(PyPetsc##TYPE##_Get);   \
+        if (PyObject_TypeCheck(src.ptr(), &PyPetsc##TYPE##_Type) == 0)  \
+          return false;                                                 \
+        value = PyPetsc##TYPE##_Get(src.ptr());                         \
+        return true;                                                    \
+      }                                                                 \
+                                                                        \
+      static handle cast(TYPE src, pybind11::return_value_policy policy, handle parent) \
+      {                                                                 \
+        VERIFY_PETSC4PY(PyPetsc##TYPE##_New);                           \
+        return pybind11::handle(PyPetsc##TYPE##_New(src));              \
+      }                                                                 \
+                                                                        \
+      operator TYPE()                                                   \
+      { return value; }                                                 \
+    }
+#else
+#define PETSC_CASTER_MACRO(TYPE, NAME)          \
+  template <> class type_caster<_p_##TYPE>      \
+    {                                           \
+    public:                                     \
+      PYBIND11_TYPE_CASTER(TYPE, _(#NAME));     \
+      bool load(handle src, bool)               \
+      {                                         \
+        throw std::runtime_error("DOLFIN has not been configured with petsc4py. Accessing underlying PETSc object requires petsc4py"); \
+        return false;                                                   \
+      }                                                                 \
+                                                                        \
+      static handle cast(TYPE src, pybind11::return_value_policy policy, handle parent) \
+      {                                                                 \
+        throw std::runtime_error("DOLFIN has not been configured with petsc4py. Accessing underlying PETSc object requires petsc4py"); \
+        return handle();                                                \
+      }                                                                 \
+                                                                        \
+      operator TYPE()                                                   \
+      { return value; }                                                 \
+    }
 #endif
 
-
-// pybind11 casters for PETSc/petsc4py objects
-
-// FIXME: Create a macro for casting code
 
 namespace pybind11
 {
   namespace detail
   {
-    template <> class type_caster<_p_Vec>
-    {
-    public:
-      PYBIND11_TYPE_CASTER(Vec, _("vec"));
-
-      // Python to C++
-      bool load(handle src, bool)
-      {
-        // FIXME: check reference counting
-        //std::cout << "Py to c++" << std::endl;
-        #ifdef HAS_PYBIND11_PETSC4PY
-        value = PyPetscVec_Get(src.ptr());
-        return true;
-        #else
-        throw std::runtime_error("DOLFIN has not been configured with petsc4py. Accessing underlying PETSc object requires petsc4py");
-        return false;
-        #endif
-      }
-
-      // C++ to Python
-      static handle cast(Vec src, pybind11::return_value_policy policy, handle parent)
-      {
-        // FIXME: check reference counting
-        #ifdef HAS_PYBIND11_PETSC4PY
-        std::cout << "C++ to Python" << std::endl;
-        return pybind11::handle(PyPetscVec_New(src));
-        #else
-        throw std::runtime_error("DOLFIN has not been configured with petsc4py. Accessing underlying PETSc object requires petsc4py");
-        return handle();
-        #endif
-      }
-
-      operator Vec()
-      { return value; }
-    };
-
-    template <> class type_caster<_p_Mat>
-    {
-    public:
-      PYBIND11_TYPE_CASTER(Mat, _("mat"));
-
-      // Python to C++
-      bool load(handle src, bool)
-      {
-        // FIXME: check reference counting
-        //std::cout << "Py to c++" << std::endl;
-        #ifdef HAS_PYBIND11_PETSC4PY
-        value = PyPetscMat_Get(src.ptr());
-        return true;
-        #else
-        throw std::runtime_error("DOLFIN has not been configured with petsc4py. Accessing underlying PETSc object requires petsc4py");
-        return false;
-        #endif
-      }
-
-      // C++ to Python
-      static handle cast(Mat src, pybind11::return_value_policy policy, handle parent)
-      {
-        // FIXME: check reference counting
-        #ifdef HAS_PYBIND11_PETSC4PY
-        std::cout << "C++ to Python" << std::endl;
-        return pybind11::handle(PyPetscMat_New(src));
-        #else
-        throw std::runtime_error("DOLFIN has not been configured with petsc4py. Accessing underlying PETSc object requires petsc4py");
-        return handle();
-        #endif
-      }
-
-      operator Mat()
-      { return value; }
-    };
-
-    template <> class type_caster<_p_KSP>
-    {
-    public:
-      PYBIND11_TYPE_CASTER(KSP, _("ksp"));
-
-      // Python to C++
-      bool load(handle src, bool)
-      {
-        // FIXME: check reference counting
-        //std::cout << "Py to c++" << std::endl;
-        #ifdef HAS_PYBIND11_PETSC4PY
-        value = PyPetscKSP_Get(src.ptr());
-        return true;
-        #else
-        throw std::runtime_error("DOLFIN has not been configured with petsc4py. Accessing underlying PETSc object requires petsc4py");
-        return false;
-        #endif
-      }
-
-      // C++ to Python
-      static handle cast(KSP src, pybind11::return_value_policy policy, handle parent)
-      {
-        // FIXME: check reference counting
-        #ifdef HAS_PYBIND11_PETSC4PY
-        //std::cout << "C++ to Python" << std::endl;
-        return pybind11::handle(PyPetscKSP_New(src));
-        #else
-        throw std::runtime_error("DOLFIN has not been configured with petsc4py. Accessing underlying PETSc object requires petsc4py");
-        return handle();
-        #endif
-      }
-
-      operator KSP()
-      { return value; }
-    };
-
-    template <> class type_caster<_p_DM>
-    {
-    public:
-      PYBIND11_TYPE_CASTER(DM, _("DM"));
-
-      // Python to C++
-      bool load(handle src, bool)
-      {
-        // FIXME: check reference counting
-        #ifdef HAS_PYBIND11_PETSC4PY
-        //std::cout << "Py to C++ (DM)" << std::endl;
-        value = PyPetscDM_Get(src.ptr());
-        //std::cout << "Returning" << std::endl;
-        return true;
-        #else
-        throw std::runtime_error("DOLFIN has not been configured with petsc4py. Accessing underlying PETSc object requires petsc4py");
-        return false;
-        #endif
-
-      }
-
-      // C++ to Python
-      static handle cast(DM src, pybind11::return_value_policy policy, handle parent)
-      {
-        // FIXME: check reference counting
-        #ifdef HAS_PYBIND11_PETSC4PY
-        //std::cout << "C++ to Python (DM)" << std::endl;
-        return pybind11::handle(PyPetscDM_New(src));
-        #else
-        throw std::runtime_error("DOLFIN has not been configured with petsc4py. Accessing underlying PETSc object requires petsc4py");
-        return pybind11::handle();
-        #endif
-      }
-
-      operator DM()
-      { return value; }
-    };
+    PETSC_CASTER_MACRO(Vec, vec);
+    PETSC_CASTER_MACRO(Mat, mat);
+    PETSC_CASTER_MACRO(KSP, ksp);
+    PETSC_CASTER_MACRO(DM, dm);
   }
 }
+
+#undef PETSC_CASTER_MACRO
 
 #endif
 #endif