+++ test/unit/python/fem/test_system_assembler.py
@@ -129,7 +129,7 @@ def test_facet_assembly():
 
         # Define normal component, mesh size and right-hand side
         n = FacetNormal(mesh)
-        h = CellSize(mesh)
+        h = 2*Circumradius(mesh)
         h_avg = (h('+') + h('-'))/2
         f = Expression("500.0*exp(-(pow(x[0] - 0.5, 2) + pow(x[1] - 0.5, 2)) / 0.02)", degree=1)